Refactors WP data exporter interface into own class
